fix: 유저 저장 및 조회 기능 DB에 반영 안되는 이슈 해결

+++ src/main/java/swe4/saju_taro/repository/UserRepository.java
@@ -5,9 +5,8 @@ import swe4.saju_taro.domain.User;
 
 import java.util.Optional;
 
-public interface UserRepository{
-    // 자동 메서드
-    Optional<User> findById(Long id);
+public interface UserRepository {
+    Optional<User> findById(Integer id);
 
     User save(User user);
 }
